bugfix: fix compilation, GPU decode yuv420p, builder, README
- Fix LogLevel typedef in logutil.h (was inside #ifdef)
- Move gpu_backend.c/libav_decoder.c to CORE_SOURCES for CPU builds
- Rewrite corrupted build_cuda.bat with prerequisite checks
- Fix fused setlocal/echo in build_verbose.bat
- Use yuv420p output for h264/hevc for NVDEC compatibility
- Add hevc/libx265 codec support (yuv420p, CRF 10 ultrafast)
- Add threadpool_submit_batch for reduced lock contention
- Remove redundant memset in frame generation
- Auto-detect CPU core count for worker threads
- Update README: benchmarks, codec guide, MSVC CUDA note
- Add logutil.c/h to README repository structure

+++ src/simd_decode.c
@@ -458,8 +458,9 @@ void tile_expand_bits(const uint8_t *src_bits,
             int bit_idx = row * grid_cols + col;
             uint8_t val = src_bits[bit_idx] ? 255 : 0;
             int x = block_x + col * block_size;
+            /* Use memset for bulk fill — block_size is small (8 or 16) */
             for (int yy = y; yy < y + block_size; ++yy)
-                memset(frame + yy * stride + x, val, (size_t)block_size);
+                memset(frame + (size_t)yy * (size_t)stride + (size_t)x, val, (size_t)block_size);
         }
     }
 }
